ログインユーザの削除を行えないように改修 それに伴いテストケースを追加

+++ src/main/java/com/ne/jp/shumipro/controller/AdminController.java
@@ -2,6 +2,7 @@ package com.ne.jp.shumipro.controller;
 
 import java.util.List;
 
+import com.ne.jp.shumipro.component.SessionData;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -18,22 +19,28 @@ public class AdminController {
 
 	@Autowired
 	AdminService adminService;
+
+	@Autowired
+	SessionData sessionData;
 	
 	@RequestMapping("")
 	public String admin(Model model) {
 		
 		List<UserAuth> userAuthList = adminService.getUsersAll();
 		model.addAttribute("userList", userAuthList);
-		
+		model.addAttribute("loginUsername", sessionData.getUsername());
 		return "admin/admin";
 	}
 	
 	@RequestMapping("/delete")
 	public String delete(RedirectAttributes redirectAttrs, Model model, @RequestParam("username") String username) {
-		
-		adminService.deleteUser(username);
-		redirectAttrs.addFlashAttribute("message", "削除が完了しました");
-		
+		if (username == null || sessionData.getUsername().equals(username)){
+			redirectAttrs.addFlashAttribute("message", "削除に失敗しました");
+		} else {
+			adminService.deleteUser(username);
+			redirectAttrs.addFlashAttribute("message", "削除が完了しました");
+		}
+
 		return "redirect:/admin";
 	}
 }
